Python-text-to-speech-converter/requirements.txt
newspaper3k
gtts
nltk
beautifulsoup4
lxml
requests